change WORD to size_t (issue for inputs > 4G)

+++ src/sha256.c
@@ -110,7 +110,7 @@ static void sha256_init(SHA256_CTX *ctx)
 
 static void sha256_update(SHA256_CTX *ctx, const BYTE data[], size_t len)
 {
-	WORD i;
+	size_t i;
 
 	for (i = 0; i < len; ++i) {
 		ctx->data[ctx->datalen] = data[i];
